+++ score-lib/src/main/java/network/balanced/score/lib/interfaces/Loans.java
@@ -113,7 +113,7 @@ public interface Loans extends Name, TokenFallback, AdminAddress, DexAddress, Go
     void toggleAssetActive(String _symbol);
 
     @External
-    boolean precompute(BigInteger _snapshot_id, int batch_size);
+    boolean precompute(BigInteger _snapshot_id, BigInteger batch_size);
 
     @External(readonly = true)
     BigInteger getTotalValue(String _name, BigInteger _snapshot_id);
